% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPdata.R
\docType{data}
\name{GPdata}
\alias{GPdata}
\title{GPdata}
\format{A data frame with 355 rows and 2 variables:
\describe{
  \item{\code{gem}}{Dutch municipality}
  \item{\code{prov}}{Corresponding province}
}}
\usage{
GPdata
}
\description{
Dataset indicating in which province each municipality is located.
}
\keyword{datasets}
