% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLDgem.R
\name{NLDgem}
\alias{NLDgem}
\title{NLDgem}
\usage{
NLDgem(
  stat = NULL,
  varname = "value",
  getGEM = FALSE,
  title = waiver(),
  subtitle = NULL,
  copyright = NULL,
  mincol = "turquoise1",
  maxcol = "steelblue4",
  legendposition = c(0.05, 0.75),
  theme_add = theme(),
  na.color = "grey",
  legend.breaks = waiver(),
  prov.seperate = NULL
)
}
\arguments{
\item{stat}{Dataset containing the names of all Dutch municipalities in the
first column and the values of the variables of interest in the second column.
If no dataset is provided, municipalities will be coloured randomly.}

\item{varname}{Name of the variable of interest as shown above the legend.}

\item{getGEM}{If TRUE the functions returns a vector with all Dutch municipalities.}

\item{title, subtitle}{Title and subtitle of the plot. NULL if omitted.}

\item{copyright}{Adds a footnote with copyright sign and the string provided to the plot}

\item{mincol, maxcol}{Minimum and maximum values of the continuous colour scale}

\item{legendposition}{Place of the legend. Needs a vector of two numbers between
0 and 1, where (0,0) is topright and (1,1) bottom left.}

\item{theme_add}{Change the theme of the plot by adding a theme() element.}
}
\description{
Generates a plot of the Netherlands where each municipality is
coloured according to some value, using a continuous scale.
}
\examples{
NLDgem(stat = NULL, varname = 'random')
}
