% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLDprov.R
\name{NLDprov}
\alias{NLDprov}
\title{NLDprov}
\usage{
NLDprov(
  stat = NULL,
  varname = "value",
  getPROV = FALSE,
  title = "Kaart van Nederland",
  subtitle = NULL,
  copyright = NULL,
  mincol = "turquoise1",
  maxcol = "steelblue4",
  na.color = "grey",
  legendposition = c(0.05, 0.75),
  theme_add = theme(),
  legend.breaks = waiver()
)
}
\arguments{
\item{stat}{Dataset containing the names of all Dutch provinces in the
first column and the values of the variables of interest in the second column.
If no dataset is provided, provinces will be coloured randomly.}

\item{varname}{Name of the variable of interest as shown above the legend.}

\item{title, subtitle}{Title and subtitle of the plot. NULL if omitted.}

\item{copyright}{Adds a footnote with copyright sign and the string provided to the plot}

\item{mincol, maxcol}{Minimum and maximum values of the continuous colour scale}

\item{legendposition}{Place of the legend. Needs a vector of two numbers between
0 and 1, where (0,0) is topright and (1,1) bottom left.}

\item{theme_add}{Change the theme of the plot by adding a theme() element.}

\item{getGEM}{If TRUE the functions returns a vector with all Dutch provinces}
}
\description{
Generates a plot of the Netherlands where each province is
coloured according to some value, using a continuous scale.
}
\examples{
NLDprov(stat = NULL, varname = 'random')
}
