% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_here.R
\name{setwd_here}
\alias{setwd_here}
\title{setwd_here()}
\usage{
setwd_here()
}
\description{
Shortcut for setting the working directory  to the map where the
current file is located.
}
\examples{
setwd_here()
}
