% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_newgrey.R
\name{theme_newgrey}
\alias{theme_newgrey}
\title{theme_newgrey}
\usage{
theme_newgrey(base_size = 12, ...)
}
\arguments{
\item{base_size}{Size of standard text, 12 by default.}

\item{...}{Other variables that are allowed within 'theme' environment.}
}
\description{
Theme for ggplot
}
