% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_notepad.R
\name{theme_notepad}
\alias{theme_notepad}
\title{theme_notepad}
\usage{
theme_notepad(
  base_size = 12,
  font = "mono",
  bg.colour = "#fffef7",
  void = FALSE,
  ...
)
}
\arguments{
\item{base_size}{Size of standard text, 12 by default.}

\item{font}{Font used in theme, 'mono' by default.}

\item{bg.colour}{Background color used in theme.}

\item{void}{Boolean variable indiciting whether grid is visible or not.}

\item{...}{Other variables that are allowed within 'theme' environment.}
}
\description{
Theme for ggplot
}
