% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean}
\usage{
clean(envir = .GlobalEnv)
}
\description{
Cleans the global environment, console and graphical device with an animation.
Purely for fun purposes.
}
\examples{
clean()
}
